% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.folder.R
\name{import.files}
\alias{import.files}
\title{output is going to be a list of
   numbers of files index (integer vector)
   sample ID given in the sequence editor (string vector)
   sample type (i.e. Standard, Sample or Blank) (factor vector)
   list of raw data values (list of dataframes)}
\usage{
import.files()
}
\description{
output is going to be a list of
   numbers of files index (integer vector)
   sample ID given in the sequence editor (string vector)
   sample type (i.e. Standard, Sample or Blank) (factor vector)
   list of raw data values (list of dataframes)
}
