% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_List.R
\name{maxList}
\alias{maxList}
\title{These function return the maximum value contained in a specific column in list of dataframes}
\usage{
maxList(l, col, ...)
}
\description{
These function return the maximum value contained in a specific column in list of dataframes
}
