% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_correction.R
\name{mb_df}
\alias{mb_df}
\title{methods 
    exponential law   -> "exp"
    power law         -> "pow"
    linear law        -> "lin"}
\usage{
mb_df(df, method = "exp", ref.ratio = c("146Nd", "144Nd"))
}
\description{
methods 
    exponential law   -> "exp"
    power law         -> "pow"
    linear law        -> "lin"
}
